% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.predictNMBsim}
\alias{plot.predictNMBsim}
\title{Create plots of from predictNMB simulations.}
\usage{
\method{plot}{predictNMBsim}(
  x,
  what = c("nmb", "inb", "cutpoints"),
  inb_ref_col = NA,
  ci = 0.95,
  methods_order = NULL,
  n_bins = 40,
  label_wrap_width = 12,
  fill_cols = c("grey50", "#ADD8E6"),
  median_line_size = 2,
  median_line_alpha = 0.5,
  median_line_col = "black",
  rename_vector,
  extra_theme = ggplot2::theme(panel.spacing = ggplot2::unit(0, "lines"), axis.ticks.x =
    ggplot2::element_blank(), axis.text.x = ggplot2::element_blank(), strip.background =
    ggplot2::element_rect(fill = "#f1f1f1"), panel.grid.minor = ggplot2::element_blank()),
  ...
)
}
\arguments{
\item{x}{a \code{predictNMBsim} object.}

\item{what}{what to summarise: one of "nmb", "inb" or "cutpoints". Defaults to "nmb".}

\item{inb_ref_col}{which cutpoint method to use as the reference strategy when calculating the incremental net monetary benefit. See \code{do_nmb_sim} for more information.}

\item{ci}{the confidence level. Defaults to 0.95 (coloured area of distribution represents 95\% CIs).}

\item{methods_order}{the order (left to right) to display the cutpoint methods.}

\item{n_bins}{the number of bins used when constructing histograms. Defaults to 40.}

\item{label_wrap_width}{the number of characters in facet labels at which the label is wrapped. Default is 12.}

\item{fill_cols}{vector containing the colours used for fill aesthetic of histograms. The first colour represents the area outside of the confidence region, second colour shows the confidence region. Defaults to \code{c("grey50", "#ADD8E6")}.}

\item{median_line_size}{size of line used to represent the median of distribution. Defaults to 2.}

\item{median_line_alpha}{alpha for line used to represent the median of distribution. Defaults to 0.5.}

\item{median_line_col}{colour of line used to represent the median of distribution. Defaults to \code{"black"}.}

\item{rename_vector}{a named vector for renaming the methods in the summary. The values of the vector are the default names and the names given are the desired names in the output.}

\item{extra_theme}{additional theme applied to plot. Defaults to some tasteful changes to reduce clutter in the plot.}

\item{...}{additional (unused) arguments.}
}
\value{
Returns a \code{ggplot} object
}
\description{
Create plots of from predictNMB simulations.
}
\details{
This plot method works with \code{predictNMBsim} objects that are created using \code{do_nmb_sim()}. Can be used to visualise distributions from simulations for different cutpoint methods.
}
\examples{

if (FALSE) {
  get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
  sim_screen_obj <- screen_simulation_inputs(
    n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1), event_rate = c(0.1, 0.2, 0.3),
    fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
  )

  plot(sim_screen_obj)
  plot(sim_screen_obj, x_axis_var = "event_rate", constants = c(sim_auc = 0.8), dodge_width=0.02)
}
}
