% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNMB-package.R
\docType{package}
\name{predictNMB-package}
\alias{predictNMB}
\alias{predictNMB-package}
\title{predictNMB: Evaluate Clinical Prediction Models by Net Monetary Benefit}
\description{
To learn more about how to use predictNMB, see the vignettes at the
\href{https://rwparsons.github.io/predictNMB/index.html}{website} or using
the following code:

\code{browseVignettes(package = "predictNMB")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rwparsons.github.io/predictNMB/}
  \item Report bugs at \url{https://github.com/rwparsons/predictNMB/issues}
}

}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons94@gmail.com} (\href{https://orcid.org/0000-0002-6053-8174}{ORCID})

Authors:
\itemize{
  \item Robin Blythe (\href{https://orcid.org/0000-0002-3643-4332}{ORCID})
  \item Adrian Barnett (\href{https://orcid.org/0000-0001-6339-0374}{ORCID})
}

Other contributors:
\itemize{
  \item Susanna Cramb (\href{https://orcid.org/0000-0001-9041-9531}{ORCID}) [contributor]
  \item Steven McPhail (\href{https://orcid.org/0000-0002-1463-662X}{ORCID}) [contributor]
}

}
\keyword{internal}
