% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_simulation_inputs.R
\name{print.predictNMBscreen}
\alias{print.predictNMBscreen}
\title{Print a summary of a predictNMBscreen object}
\usage{
\method{print}{predictNMBscreen}(x, ...)
}
\arguments{
\item{x}{A \code{predictNMBscreen} object.}

\item{...}{Optional, ignored arguments.}
}
\value{
\code{print(x)} returns \code{x} invisibly.
}
\description{
Print a summary of a predictNMBscreen object
}
\examples{
\donttest{
get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
sim_screen_obj <- screen_simulation_inputs(
  n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1),
  event_rate = 0.1,
  fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
)
print(sim_screen_obj)
}
}
