% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{theme_sim}
\alias{theme_sim}
\title{Returns a \code{ggplot2} theme that reduces clutter in an \code{autoplot()}
of a \code{predictNMBsim} object.}
\usage{
theme_sim()
}
\value{
Returns a \code{ggplot2} theme.
}
\description{
Returns a \code{ggplot2} theme that reduces clutter in an \code{autoplot()}
of a \code{predictNMBsim} object.
}
\examples{
\donttest{
get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
sim_obj <- do_nmb_sim(
  sample_size = 200, n_sims = 50, n_valid = 10000, sim_auc = 0.7,
  event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
)

autoplot(sim_obj) + theme_sim()
}

}
