% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nmb_sim.R
\name{do_nmb_sim}
\alias{do_nmb_sim}
\title{Do the predictNMB simulation, evaluating the net monetary benefit of the simulated model.}
\usage{
do_nmb_sim(
  sample_size,
  n_sims,
  n_valid,
  sim_auc,
  event_rate,
  cutpoint_methods = get_inbuilt_cutpoint(return_all_methods = TRUE),
  fx_nmb_training,
  fx_nmb_evaluation,
  meet_min_events = TRUE
)
}
\arguments{
\item{sample_size}{Sample size of training set. If missing, a sample size calculation will be performed and the calculated size will be used.}

\item{n_sims}{Number of simulations to run.}

\item{n_valid}{Sample size for evaluation set.}

\item{sim_auc}{Simulated model discrimination (AUC).}

\item{event_rate}{Simulated event rate of the binary outcome being predicted.}

\item{cutpoint_methods}{cutpoint methods to include. Defaults to use the inbuilt methods.}

\item{fx_nmb_training}{Function that returns named vector of NMB assigned to classifications use for obtaining cutpoint on training set}

\item{fx_nmb_evaluation}{Function that returns named vector of NMB assigned to classifications use for obtaining cutpoint on evaluation set}

\item{meet_min_events}{Whether or not to incrementally add samples until the expected number of events (\code{sample_size * event_rate}) is met. (Applies to sampling of training data only.)}
}
\value{
predictNMBsim
}
\description{
Do the predictNMB simulation, evaluating the net monetary benefit of the simulated model.
}
\examples{

get_nmb <- function() c("TP" = 1, "TN" = 2, "FP" = 3, "FN" = 4)
do_nmb_sim(
  n_sims = 50, n_valid = 10000, sim_auc = 0.7, event_rate = 0.1,
  fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
)
}
