% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint_methods.R
\name{get_thresholds}
\alias{get_thresholds}
\title{Gets probability thresholds given predicted probabilities, outcomes and NMB.}
\usage{
get_thresholds(predicted, actual, nmb, cutpoint_methods = NULL)
}
\arguments{
\item{predicted}{vector of predicted probabilities}

\item{actual}{vector of actual outcomes}

\item{nmb}{named vector containing NMB assigned to true positives, true negatives, false positives and false negatives}

\item{cutpoint_methods}{which cutpoint method(s) to return. The default (NULL) uses all the inbuilt methods.}
}
\value{
list
}
\description{
Gets probability thresholds given predicted probabilities, outcomes and NMB.
}
\examples{
get_thresholds(
  predicted = runif(1000),
  actual = sample(c(0, 1), size = 1000, replace = TRUE),
  nmb = c("TP" = 1, "TN" = 2, "FP" = 3, "FN" = 4)
)
}
